% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corf.R
\name{corf}
\alias{corf}
\title{Select informative PCG subset}
\usage{
corf(train.pcg, train.circ, gene.index, num = 50)
}
\arguments{
\item{train.pcg}{training protein coding dataset,
which should be a numeric matrix with row names indicating samples,
and column names indicating protein coding gene IDs.}

\item{train.circ}{training circRNA expression dataset,
which should be a numeric matrix with row names indicating samples,
and column names indicating circRNA IDs.}

\item{gene.index}{either gene name (character) or index (column number) of circRNA to be imputed.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model. Default is 50 genes.}
}
\value{
a numeric matrix, which should be a subset of protein coding genes correlated with circRNA of interest.
}
\description{
Function to filter informative protein coding genes based on correlation with circRNA expression
}
