% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_proc.R
\name{cv_proc}
\alias{cv_proc}
\title{Process cross-validation results}
\usage{
cv_proc(res)
}
\arguments{
\item{res}{The output object from imirage.cv}
}
\value{
a processed matrix containing 3 columns: Spearman's correlation coefficient, P-value and root mean
squared error from cross-validation analysis
}
\description{
This function returns useful information by organizing the output from cross-validation analysis.
Used internally by \link[ICE]{ICE_cv_loop}
}
