% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_circ.R
\name{filter_circ}
\alias{filter_circ}
\title{Pre-standardization filter to select circRNA expressed in a specific proportion of samples}
\usage{
filter_circ(train.circ, cutoff = 75, threshold = 0)
}
\arguments{
\item{train.circ}{an expression matrix with circRNA in columns, samples in rows}

\item{cutoff}{percentage of samples in which the circRNA should be expressed}

\item{threshold}{the numeric threshold defining expression of circRNA. Default threshold = 0.}
}
\value{
filtered circRNA expression matrix
}
\description{
return a subset of the circRNA expression dataset containing circRNAs that are expressed
in at least a specified percentage of samples based on a user-defined expression threshold
}
\examples{
New.mir <-  filter_circ(GA.mir, cutoff=95, threshold=0)

}
