% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE.R
\docType{package}
\name{ICE}
\alias{ICE}
\title{ICE: Impute CircRNA Expression}
\usage{
ICE(train.circ, train.pcg, new.pcg, gene.index, method = "KNN", num = 100, ...)
}
\arguments{
\item{train.circ}{training circRNA expression dataset, which should be a numeric matrix with row names indicating
samples, and column names indicating circRNA IDs.}

\item{train.pcg}{training protein coding dataset, which should be a numeric matrix with with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{new.pcg}{protein coding expression dataset to be used for prediction, which should be a numeric matrix with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{gene.index}{either gene name (character) or index (column number) of circRNA to be imputed.}

\item{method}{method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or
"SVM" for support vector machines. Uses KNN by default.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model.
Default is 100 genes.}

\item{...}{}
}
\value{

}
\examples{
data("mionco.circ")
data("mionco.pcg")
pred.circ <- ICE(train.pcg = mionco.pcg, train.circ = mionco.circ, new.pcg = mionco.pcg, gene.index = "hsa_circ_0000801")
}
