% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE_cv_entire.R
\name{ICE_cv_entire}
\alias{ICE_cv_entire}
\title{ICE cross-validation loop function for full circRNA matrix}
\usage{
ICE_cv_entire(train.pcg, train.circ, method = "KNN", ...)
}
\arguments{
\item{train.pcg}{training protein coding dataset. a numeric matrix with with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{train.circ}{training miRNA expression dataset. a numeric matrix with row names indicating
samples, and column names indicating miRNA IDs.}

\item{method}{method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or
"SVM" for support vector machines.}

\item{...}{optional parameters that can be passed on to the machine-learning method:
RF (\link[randomForest]{randomForest}), KNN (\link[FNN]{knn.reg}) or SVM(\link[e1071]{svm})}
}
\value{
a matrix containing Spearman's correlation coefficient, P-value and RMSE from the cross-validation analysis
of the complete miRNA training dataset
}
\description{
Convinient wrapper for \link[ICE]{ICE_cv} that performs cross-validation analysis for
assessing imputation accuracies for all miRNAs using the training datasets
}
\examples{
ICE_cv_entire(train.pcg = GA.pcg, train.circ = GA.mir, method = "KNN")
ICE_cv_entire(train.pcg = GA.pcg, train.circ = GA.mir, method = "SVM")

}
