\name{ICHemophiliac}
\alias{ICHemophiliac}
\docType{data}
\title{Hemophiliac Data Set}
\description{
}
\usage{data(ICHemophiliac)}
\format{
  The format is:
 num [1:96, 1:2] 10 12 9 15 5 1 15 1 9 11 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "left" "right"
}
\keyword{datasets}
