% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_mat.R
\name{match_mat}
\alias{match_mat}
\title{Match gene expression matrices by gene ID}
\usage{
match_mat(train.pcg, new.pcg)
}
\arguments{
\item{train.pcg}{Gene expression matric with gene IDs as column names}

\item{new.pcg}{Gene expression matric with gene IDs as column names}
}
\value{
A list containing two matrices with common genes
}
\description{
Returns expression matrices with common genes
}
\examples{
temp <- match_mat(GA.pcg, HS.pcg)
Matched.GA.pcg <- data.matrix(temp[[1]])
Matched.HS.pcg <- data.matrix(temp[[2]])

}
