% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE_cv.R
\name{ICE_cv}
\alias{ICE_cv}
\title{Cross validation function for ICE imputation accuracy}
\usage{
ICE_cv(
  train.pcg,
  train.circ,
  gene.index,
  num = 50,
  method = "KNN",
  folds = 10,
  filter = TRUE,
  cor.method = "spearman",
  ...
)
}
\arguments{
\item{train.pcg}{training protein coding dataset, which should be a numeric matrix with with row names indicating
samples, and column names indicating protein coding gene IDs.}

\item{train.circ}{training circRNA expression dataset, which should be a numeric matrix with row names indicating
samples, and column names indicating circRNA IDs.}

\item{gene.index}{either gene name (character) or index (column number) of circRNA to be imputed.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model.
Default is 100 genes.}

\item{method}{method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or
"SVM" for support vector machines.}

\item{folds}{number specifying folds (k) of cross validation to obtain imputation accuracy.
Default is k=10.}

\item{filter}{if training datasets should be filtered? Default is TRUE, but should be set to FALSE when called in loop.}

\item{cor.method}{a character string indicating
which correlation coefficient is to be used for the correlation test.
Default is "spearman", could also be "pearson".}

\item{...}{optional parameters that can be passed on to the machine-learning functions
RF (\link[randomForest]{randomForest}), KNN (\link[FNN]{knn.reg}) or SVM(\link[e1071]{svm})}
}
\value{
a matrix with three values corresponding to correlation coefficient,
P-value of the fit and root mean squared error (RMSE).
}
\description{
Function to obtain accuracy parameters: correlation coefficient, P-value and RMSE of
imputation model
}
\examples{
data(iMIRAGE.datasets)
ICE_cv(GA.pcg, GA.mir, gene.index="hsa-let-7c", method="KNN", num=100)
ICE_cv(GA.pcg, GA.mir, gene.index=25, method="KNN", num=100)

}
