\name{ICHemophiliac}
\alias{ICHemophiliac}
\docType{data}
\title{Hemophiliac Data Set}
\description{Time (interval-censored) to onset of HIV infection.
}
\usage{data(ICHemophiliac)}
\format{a matrix with 2 columns.  The first column contains
the left endpoints of the intervals, and the second column
contains the right endpoints.
}
\keyword{datasets}
